pyproject.toml
[project]
name = "rag-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "pydantic>=2.12.5",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "wikipedia>=1.4.0",
]
